docshell/test/iframesandbox/test_other_auxiliary_navigation_by_location.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=785310
html5 sandboxed iframe should not be able to perform top navigation with scripts allowed
-->
<head>
<meta charset="utf-8">
<title>Test for Bug 785310 - iframe sandbox other auxiliary navigation by location tests</title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

<script>
  SimpleTest.waitForExplicitFinish();

  function runScriptNavigationTest(testCase) {
    window.onmessage = function(event) {
      if (event.data != "otherWindow") {
        ok(false, "event.data: got '" + event.data + "', expected 'otherWindow'");
      }
      ok(false, testCase.desc, "auxiliary navigation was NOT blocked");
      runNextTest();
    };
    try {
      window["testIframe"].eval(testCase.script);
    } catch(e) {
      ok(true, testCase.desc, e.message);
      runNextTest();
    }
  }

  var testCaseIndex = -1;
  testCases = [
    {
      desc: "Test 1: location.replace on auxiliary NOT opened by us should be blocked",
      script: "parent.openedWindow.location.replace('file_other_auxiliary_navigation_by_location.html')"
    },
    {
      desc: "Test 2: location.assign on auxiliary NOT opened by us should be blocked",
      script: "parent.openedWindow.location.assign('file_other_auxiliary_navigation_by_location.html')"
    },
    {
      desc: "Test 3: location.href on auxiliary NOT opened by us should be blocked",
      script: "parent.openedWindow.location.href = 'file_other_auxiliary_navigation_by_location.html'"
    },
    {
      desc: "Test 4: location.hash on auxiliary NOT opened by us should be blocked",
      script: "parent.openedWindow.location.hash = 'wibble'"
    }
  ];

  function runNextTest() {
    ++testCaseIndex;
    if (testCaseIndex == testCases.length) {
      openedWindow.close();
      SimpleTest.finish();
      return;
    }

    runScriptNavigationTest(testCases[testCaseIndex]);
  }

  window.onmessage = runNextTest;

  window.onload = function() {
    window.openedWindow = window.open("file_other_auxiliary_navigation_by_location.html", "otherWindow");
  }
</script>

</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=785310">Mozilla Bug 785310</a>
<p id="display"></p>
<div id="content">
Tests for Bug 785310
</div>

<iframe name="testIframe" sandbox="allow-scripts allow-same-origin allow-top-navigation allow-popups"></iframe>
</body>
</html>
